Split config.d.ts
