+++ package.json
@@ -46,6 +46,7 @@
     "sanitize-filename": "^1.6.2"
   },
   "devDependencies": {
+    "@types/express": "^4.17.2",
     "@types/redis": "^2.8.14",
     "coveralls": "^3.0.6",
     "eslint": "^5.16.0",
